+++ app/Http/Controllers/RequestView.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Events\RequestError;
 use App\Option\Get;
 use App\Option\Post;
 use App\Response\Header\Header;
@@ -12,8 +11,6 @@ use App\Models\RequestLog;
 use App\Models\Transformers\RequestErrorLog as RequestErrorLogTransformer;
 use App\Models\Transformers\RequestLog as RequestLogTransformer;
 use App\Response\Pagination as UtilityPagination;
-use App\Request\Validate\RequestErrorLog as RequestErrorLogValidator;
-use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Config;
 
@@ -24,7 +21,7 @@ use Illuminate\Support\Facades\Config;
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class RequestController extends Controller
+class RequestView extends Controller
 {
     protected $collection_parameters = [];
 
@@ -144,44 +141,5 @@ class RequestController extends Controller
             $get + $post,
             200
         );
-    }
-
-    /**
-     * Log a request error, these are logged when the web app receives an unexpected
-     * http status code response
-     *
-     * @return JsonResponse
-     */
-    public function createErrorLog(): JsonResponse
-    {
-        $validator = (new RequestErrorLogValidator())->create();
-        \App\Request\BodyValidation::validateAndReturnErrors($validator);
-
-        try {
-            $request_error_log = new RequestErrorLog([
-                'method' => request()->input('method'),
-                'source' => request()->input('source'),
-                'expected_status_code' => request()->input('expected_status_code'),
-                'returned_status_code' => request()->input('returned_status_code'),
-                'request_uri' => request()->input('request_uri'),
-                'debug' => request()->input('debug')
-            ]);
-            $request_error_log->save();
-
-            event(new RequestError([
-                'method' => request()->input('method'),
-                'source' => request()->input('source'),
-                'expected_status_code' => request()->input('expected_status_code'),
-                'returned_status_code' => request()->input('returned_status_code'),
-                'request_uri' => request()->input('request_uri'),
-                'referer' => request()->server('HTTP_REFERER', 'NOT SET!'),
-                'debug' => request()->input('debug')
-            ]));
-
-        } catch (Exception $e) {
-            \App\Response\Responses::failedToSaveModelForCreate();
-        }
-
-        return \App\Response\Responses::successNoContent();
     }
 }
