+++ app/Http/Controllers/ResourceView.php
@@ -14,11 +14,7 @@ use App\Response\Header\Headers;
 use App\Response\Pagination as UtilityPagination;
 use App\Models\Resource;
 use App\Models\Transformers\Resource as ResourceTransformer;
-use App\Request\Validate\Resource as ResourceValidator;
-use Exception;
-use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Config;
 
 /**
@@ -28,7 +24,7 @@ use Illuminate\Support\Facades\Config;
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceController extends Controller
+class ResourceView extends Controller
 {
     protected bool $allow_entire_collection = true;
 
@@ -234,165 +230,5 @@ class ResourceController extends Controller
             $get + $delete + $patch,
             200
         );
-    }
-
-    /**
-     * Create a new resource
-     *
-     * @param string $resource_type_id
-     *
-     * @return JsonResponse
-     */
-    public function create(string $resource_type_id): JsonResponse
-    {
-        Route\Validate::resourceType(
-            $resource_type_id,
-            $this->permitted_resource_types,
-            true
-        );
-
-        $cache_control = new Cache\Control(Auth::user()->id);
-        $cache_key = new Cache\Key();
-
-        $validator = (new ResourceValidator)->create(['resource_type_id' => $resource_type_id]);
-        \App\Request\BodyValidation::validateAndReturnErrors($validator);
-
-        try {
-            $resource = new Resource([
-                'resource_type_id' => $resource_type_id,
-                'name' => request()->input('name'),
-                'description' => request()->input('description'),
-                'effective_date' => request()->input('effective_date')
-            ]);
-            $resource->save();
-
-            $cache_control->clearPrivateCacheKeys([
-                $cache_key->resourceType($resource_type_id)
-            ]);
-
-            if (in_array($resource_type_id, $this->public_resource_types, true)) {
-                $cache_control->clearPublicCacheKeys([
-                    $cache_key->resourceType($resource_type_id)
-                ]);
-            }
-        } catch (Exception $e) {
-            \App\Response\Responses::failedToSaveModelForCreate();
-        }
-
-        return response()->json(
-            (new ResourceTransformer((New Resource())->instanceToArray($resource)))->asArray(),
-            201
-        );
-    }
-
-    /**
-     * Delete the requested resource
-     *
-     * @param string $resource_type_id,
-     * @param string $resource_id
-     *
-     * @return JsonResponse
-     */
-    public function delete(
-        string $resource_type_id,
-        string $resource_id
-    ): JsonResponse
-    {
-        Route\Validate::resource(
-            $resource_type_id,
-            $resource_id,
-            $this->permitted_resource_types,
-            true
-        );
-
-        $cache_control = new Cache\Control(Auth::user()->id);
-        $cache_key = new Cache\Key();
-
-        try {
-            (new Resource())->find($resource_id)->delete();
-
-            $cache_control->clearPrivateCacheKeys([
-                $cache_key->resourceType($resource_type_id)
-            ]);
-
-            if (in_array($resource_type_id, $this->public_resource_types, true)) {
-                $cache_control->clearPublicCacheKeys([
-                    $cache_key->resourceType($resource_type_id)
-                ]);
-            }
-
-            \App\Response\Responses::successNoContent();
-        } catch (QueryException $e) {
-            \App\Response\Responses::foreignKeyConstraintError();
-        } catch (Exception $e) {
-            \App\Response\Responses::notFound(trans('entities.resource'), $e);
-        }
-    }
-
-    /**
-     * Update the selected resource
-     *
-     * @param string $resource_type_id
-     * @param string $resource_id
-     *
-     * @return JsonResponse
-     */
-    public function update(
-        string $resource_type_id,
-        string $resource_id
-    ): JsonResponse
-    {
-        Route\Validate::resource(
-            $resource_type_id,
-            $resource_id,
-            $this->permitted_resource_types,
-            true
-        );
-
-        $cache_control = new Cache\Control(Auth::user()->id);
-        $cache_key = new Cache\Key();
-
-        $resource = (new Resource())->instance($resource_type_id, $resource_id);
-
-        if ($resource === null) {
-            \App\Response\Responses::failedToSelectModelForUpdateOrDelete();
-        }
-
-        \App\Request\BodyValidation::checkForEmptyPatch();
-
-        $validator = (new ResourceValidator())->update([
-            'resource_type_id' => (int)$resource_type_id,
-            'resource_id' => (int)$resource_id
-        ]);
-        \App\Request\BodyValidation::validateAndReturnErrors($validator);
-
-        \App\Request\BodyValidation::checkForInvalidFields(
-            array_merge(
-                (new Resource())->patchableFields(),
-                (new ResourceValidator())->dynamicDefinedFields()
-            )
-        );
-
-        foreach (request()->all() as $key => $value) {
-            $resource->$key = $value;
-        }
-
-        try {
-            $resource->save();
-
-            $cache_control->clearPrivateCacheKeys([
-                $cache_key->resourceType($resource_type_id)
-            ]);
-
-            if (in_array($resource_type_id, $this->public_resource_types, true)) {
-                $cache_control->clearPublicCacheKeys([
-                    $cache_key->resourceType($resource_type_id)
-                ]);
-            }
-        } catch (Exception $e) {
-            \App\Response\Responses::failedToSaveModelForUpdate();
-        }
-
-        \App\Response\Responses::successNoContent();
     }
 }
