+++ app/Http/Controllers/Summary/ItemView.php
@@ -3,42 +3,44 @@
 namespace App\Http\Controllers\Summary;
 
 use App\Http\Controllers\Controller;
-use App\ResourceTypeItem\Factory;
+use App\Item\Factory;
+use App\Models\Transformers\Summary\ItemCategory as ItemCategoryTransformer;
+use App\Models\Transformers\Summary\ItemMonth as ItemMonthTransformer;
+use App\Models\Transformers\Summary\ItemSubcategory as ItemSubcategoryTransformer;
+use App\Models\Transformers\Summary\ItemYear as ItemYearTransformer;
 use App\Option\Get;
-use App\Response\Cache;
 use App\Request\Parameter;
 use App\Request\Route;
 use App\Request\Validate\Boolean;
-use App\Models\Transformers\Summary\ResourceTypeItemCategory as ResourceTypeItemCategoryTransformer;
-use App\Models\Transformers\Summary\ResourceTypeItemMonth as ResourceTypeItemMonthTransformer;
-use App\Models\Transformers\Summary\ResourceTypeItemResource as ResourceTypeItemResourceTransformer;
-use App\Models\Transformers\Summary\ResourceTypeItemSubcategory as ResourceTypeItemSubcategoryTransformer;
-use App\Models\Transformers\Summary\ResourceTypeItemYear as ResourceTypeItemYearTransformer;
+use App\Response\Cache;
 use App\Response\Header\Headers;
 use Illuminate\Http\JsonResponse;
 
 /**
- * Summary for resource type items route
+ * Summary for the items route
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceTypeItemController extends Controller
+class ItemView extends Controller
 {
     private $model;
 
     /**
-     * Return the TCO for all the resources within the resource type
+     * Return the TCO for the resource or pass the request off to the relevant
+     * method
      *
      * @param string $resource_type_id
+     * @param string $resource_id
      *
      * @return JsonResponse
      */
-    public function index(string $resource_type_id): JsonResponse
+    public function index(string $resource_type_id, string $resource_id): JsonResponse
     {
-        Route\Validate::resourceType(
+        Route\Validate::resource(
             $resource_type_id,
+            $resource_id,
             $this->permitted_resource_types
         );
 
@@ -47,10 +49,10 @@ class ResourceTypeItemController extends Controller
 
         $parameters = Parameter\Request::fetch(
             $item_interface->collectionParametersKeys(),
-            $resource_type_id
+            (int)$resource_type_id,
+            (int)$resource_id
         );
 
-        $resources = false;
         $years = false;
         $months = false;
         $categories = false;
@@ -59,11 +61,6 @@ class ResourceTypeItemController extends Controller
         $month = null;
         $category = null;
         $subcategory = null;
-
-        if (array_key_exists('resources', $parameters) === true &&
-            Boolean::convertedValue($parameters['resources']) === true) {
-            $resources = true;
-        }
 
         if (array_key_exists('years', $parameters) === true &&
             Boolean::convertedValue($parameters['years']) === true) {
@@ -86,23 +83,22 @@ class ResourceTypeItemController extends Controller
         }
 
         if (array_key_exists('year', $parameters) === true) {
-            $year = (int) $parameters['year'];
+            $year = (int)$parameters['year'];
         }
 
         if (array_key_exists('month', $parameters) === true) {
-            $month = (int) $parameters['month'];
+            $month = (int)$parameters['month'];
         }
 
         if (array_key_exists('category', $parameters) === true) {
-            $category = (int) $parameters['category'];
+            $category = (int)$parameters['category'];
         }
 
         if (array_key_exists('subcategory', $parameters) === true) {
-            $subcategory = (int) $parameters['subcategory'];
+            $subcategory = (int)$parameters['subcategory'];
         }
 
         unset(
-            $parameters['resources'],
             $parameters['years'],
             $parameters['year'],
             $parameters['months'],
@@ -123,7 +119,8 @@ class ResourceTypeItemController extends Controller
 
         if ($years === true) {
             return $this->yearsSummary(
-                $resource_type_id,
+                (int)$resource_type_id,
+                (int)$resource_id,
                 $parameters
             );
         }
@@ -136,7 +133,8 @@ class ResourceTypeItemController extends Controller
         ) {
             if ($months === true) {
                 return $this->monthsSummary(
-                    $resource_type_id,
+                    (int)$resource_type_id,
+                    (int)$resource_id,
                     $year,
                     $parameters
                 );
@@ -144,7 +142,8 @@ class ResourceTypeItemController extends Controller
 
             if ($month !== null) {
                 return $this->monthSummary(
-                    $resource_type_id,
+                    (int)$resource_type_id,
+                    (int)$resource_id,
                     $year,
                     $month,
                     $parameters
@@ -152,7 +151,8 @@ class ResourceTypeItemController extends Controller
             }
 
             return $this->yearSummary(
-                $resource_type_id,
+                (int)$resource_type_id,
+                (int)$resource_id,
                 $year,
                 $parameters
             );
@@ -160,7 +160,8 @@ class ResourceTypeItemController extends Controller
 
         if ($categories === true) {
             return $this->categoriesSummary(
-                $resource_type_id,
+                (int)$resource_type_id,
+                (int)$resource_id,
                 $parameters
             );
         }
@@ -173,7 +174,8 @@ class ResourceTypeItemController extends Controller
         ) {
             if ($subcategories === true) {
                 return $this->subcategoriesSummary(
-                    $resource_type_id,
+                    (int)$resource_type_id,
+                    (int)$resource_id,
                     $category,
                     $parameters
                 );
@@ -181,7 +183,8 @@ class ResourceTypeItemController extends Controller
 
             if ($subcategory !== null) {
                 return $this->subcategorySummary(
-                    $resource_type_id,
+                    (int)$resource_type_id,
+                    (int)$resource_id,
                     $category,
                     $subcategory,
                     $parameters
@@ -189,43 +192,85 @@ class ResourceTypeItemController extends Controller
             }
 
             return $this->categorySummary(
-                $resource_type_id,
+                (int)$resource_type_id,
+                (int)$resource_id,
                 $category,
                 $parameters
             );
         }
 
-        if ($resources === true) {
-            return $this->resourcesSummary(
-                $resource_type_id,
-                $parameters
-            );
-        }
-
         if (
-            $category === true ||
-            $subcategory === true ||
-            $year === true ||
-            $month === true ||
+            $category !== null ||
+            $subcategory !== null ||
+            $year !== null ||
+            $month !== null ||
             count($search_parameters) > 0 ||
             count($filter_parameters) > 0
         ) {
             return $this->filteredSummary(
-                $resource_type_id,
+                (int)$resource_type_id,
+                (int)$resource_id,
                 $category,
                 $subcategory,
                 $year,
                 $month,
                 $parameters,
                 (count($search_parameters) > 0 ? $search_parameters : []),
-                (count($filter_parameters) > 0 ? $filter_parameters : []),
+                (count($filter_parameters) > 0 ? $filter_parameters : [])
             );
         }
 
-        return $this->summary(
-            $resource_type_id,
+        return $this->tcoSummary(
+            (int)$resource_type_id,
+            (int)$resource_id,
             $parameters
         );
+    }
+
+    /**
+     * Return the annualised summary for a resource
+     *
+     * @param int $resource_type_id
+     * @param int $resource_id
+     * @param array $parameters
+     *
+     * @return JsonResponse
+     */
+    private function yearsSummary(
+        int $resource_type_id,
+        int $resource_id,
+        array $parameters
+    ): JsonResponse {
+        $cache_control = new Cache\Control($this->user_id);
+        $cache_control->setTtlOneWeek();
+
+        $cache_summary = new Cache\Summary();
+        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
+
+        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
+
+            $summary = $this->model->yearsSummary(
+                $resource_type_id,
+                $resource_id,
+                $parameters
+            );
+
+            $collection = array_map(
+                static function ($year) {
+                    return (new ItemYearTransformer($year))->asArray();
+                },
+                $summary
+            );
+
+            $this->assignContentToCache(
+                $summary,
+                $collection,
+                $cache_control,
+                $cache_summary
+            );
+        }
+
+        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
     }
 
     private function assignContentToCache(
@@ -240,7 +285,6 @@ class ResourceTypeItemController extends Controller
             addParameters(Parameter\Request::xHeader())->
             addFilters(Parameter\Filter::xHeader())->
             addSearch(Parameter\Search::xHeader());
-
 
         if (array_key_exists(0, $summary)) {
             if (array_key_exists('last_updated', $summary[0]) === true) {
@@ -258,192 +302,10 @@ class ResourceTypeItemController extends Controller
     }
 
     /**
-     * Return the total summary for all the resources in the resource type
+     * Return the monthly summary for a specific year
      *
      * @param int $resource_type_id
-     * @param array $parameters
-     *
-     * @return JsonResponse
-     */
-    private function summary(
-        int $resource_type_id,
-        array $parameters
-    ): JsonResponse
-    {
-        $cache_control = new Cache\Control($this->user_id);
-        $cache_control->setTtlOneWeek();
-
-        $cache_summary = new Cache\Summary();
-        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
-
-        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
-
-            $summary = $this->model->summary(
-                $resource_type_id,
-                $parameters
-            );
-
-            $total = '0.00';
-            if (array_key_exists(0, $summary) && array_key_exists('total', $summary[0])) {
-                $total = number_format($summary[0]['total'], 2, '.', '');
-            }
-
-            $collection = [
-                'total' => $total
-            ];
-
-            $this->assignContentToCache(
-                $summary,
-                $collection,
-                $cache_control,
-                $cache_summary
-            );
-        }
-
-        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
-    }
-
-    /**
-     * Return the total summary for all the resources in the resource type
-     * grouped by resource
-     *
-     * @param int $resource_type_id
-     * @param array $parameters
-     *
-     * @return JsonResponse
-     */
-    private function resourcesSummary(
-        int $resource_type_id,
-        array $parameters
-    ): JsonResponse
-    {
-        $cache_control = new Cache\Control($this->user_id);
-        $cache_control->setTtlOneWeek();
-
-        $cache_summary = new Cache\Summary();
-        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
-
-        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
-
-            $summary = $this->model->resourcesSummary(
-                $resource_type_id,
-                $parameters
-            );
-
-            $collection = array_map(
-                static function ($resource) {
-                    return (new ResourceTypeItemResourceTransformer($resource))->asArray();
-                },
-                $summary
-            );
-
-            $this->assignContentToCache(
-                $summary,
-                $collection,
-                $cache_control,
-                $cache_summary
-            );
-        }
-
-        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
-    }
-
-    /**
-     * Return the total summary for all the resources in the resource type
-     * grouped by year
-     *
-     * @param int $resource_type_id
-     * @param array $parameters
-     *
-     * @return JsonResponse
-     */
-    private function yearsSummary(
-        int $resource_type_id,
-        array $parameters
-    ): JsonResponse
-    {
-        $cache_control = new Cache\Control($this->user_id);
-        $cache_control->setTtlOneWeek();
-
-        $cache_summary = new Cache\Summary();
-        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
-
-        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
-
-            $summary = $this->model->yearsSummary(
-                $resource_type_id,
-                $parameters
-            );
-
-            $collection = array_map(
-                static function ($year) {
-                    return (new ResourceTypeItemYearTransformer($year))->asArray();
-                },
-                $summary
-            );
-
-            $this->assignContentToCache(
-                $summary,
-                $collection,
-                $cache_control,
-                $cache_summary
-            );
-        }
-
-        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
-    }
-
-    /**
-     * Return the total summary for all the resources in the resource type
-     * for the requested year
-     *
-     * @param int $resource_type_id
-     * @param int $year
-     * @param array $parameters
-     *
-     * @return JsonResponse
-     */
-    private function yearSummary(
-        int $resource_type_id,
-        int $year,
-        array $parameters
-    ): JsonResponse
-    {
-        $cache_control = new Cache\Control($this->user_id);
-        $cache_control->setTtlOneWeek();
-
-        $cache_summary = new Cache\Summary();
-        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
-
-        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
-
-            $summary = $this->model->yearSummary(
-                $resource_type_id,
-                $year,
-                $parameters
-            );
-
-            $collection = [];
-            if (array_key_exists(0, $summary)) {
-                $collection = (new ResourceTypeItemYearTransformer($summary[0]))->asArray();
-            }
-
-            $this->assignContentToCache(
-                $summary,
-                $collection,
-                $cache_control,
-                $cache_summary
-            );
-        }
-
-        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
-    }
-
-    /**
-     * Return the total summary for all the resources in the resource type
-     * grouped by year
-     *
-     * @param int $resource_type_id
+     * @param int $resource_id
      * @param int $year
      * @param array $parameters
      *
@@ -451,10 +313,10 @@ class ResourceTypeItemController extends Controller
      */
     private function monthsSummary(
         int $resource_type_id,
+        int $resource_id,
         int $year,
         array $parameters
-    ): JsonResponse
-    {
+    ): JsonResponse {
         $cache_control = new Cache\Control($this->user_id);
         $cache_control->setTtlOneWeek();
 
@@ -465,13 +327,14 @@ class ResourceTypeItemController extends Controller
 
             $summary = $this->model->monthsSummary(
                 $resource_type_id,
+                $resource_id,
                 $year,
                 $parameters
             );
 
             $collection = array_map(
                 static function ($month) {
-                    return (new ResourceTypeItemMonthTransformer($month))->asArray();
+                    return (new ItemMonthTransformer($month))->asArray();
                 },
                 $summary
             );
@@ -488,10 +351,10 @@ class ResourceTypeItemController extends Controller
     }
 
     /**
-     * Return the total summary for all the resources in the resource type
-     * for a specific month
+     * Return the month summary for a specific year and month
      *
      * @param int $resource_type_id
+     * @param int $resource_id
      * @param int $year
      * @param int $month
      * @param array $parameters
@@ -500,11 +363,11 @@ class ResourceTypeItemController extends Controller
      */
     private function monthSummary(
         int $resource_type_id,
+        int $resource_id,
         int $year,
         int $month,
         array $parameters
-    ): JsonResponse
-    {
+    ): JsonResponse {
         $cache_control = new Cache\Control($this->user_id);
         $cache_control->setTtlOneWeek();
 
@@ -515,6 +378,7 @@ class ResourceTypeItemController extends Controller
 
             $summary = $this->model->monthSummary(
                 $resource_type_id,
+                $resource_id,
                 $year,
                 $month,
                 $parameters
@@ -522,7 +386,7 @@ class ResourceTypeItemController extends Controller
 
             $collection = [];
             if (array_key_exists(0, $summary)) {
-                $collection = (new ResourceTypeItemMonthTransformer($summary[0]))->asArray();
+                $collection = (new ItemMonthTransformer($summary[0]))->asArray();
             }
 
             $this->assignContentToCache(
@@ -537,19 +401,21 @@ class ResourceTypeItemController extends Controller
     }
 
     /**
-     * Return the total summary for all the resources in the resource type
-     * grouped by category
+     * Return the total cost for a specific year
      *
-     * @param int $resource_type_id
+     * @param int $resource_type_id ,
+     * @param int $resource_id
+     * @param int $year
      * @param array $parameters
      *
      * @return JsonResponse
      */
-    private function categoriesSummary(
+    private function yearSummary(
         int $resource_type_id,
+        int $resource_id,
+        int $year,
         array $parameters
-    ): JsonResponse
-    {
+    ): JsonResponse {
         $cache_control = new Cache\Control($this->user_id);
         $cache_control->setTtlOneWeek();
 
@@ -558,14 +424,59 @@ class ResourceTypeItemController extends Controller
 
         if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
 
+            $summary = $this->model->yearSummary(
+                $resource_type_id,
+                $resource_id,
+                $year,
+                $parameters
+            );
+
+            $collection = [];
+            if (array_key_exists(0, $summary)) {
+                $collection = (new ItemYearTransformer($summary[0]))->asArray();
+            }
+
+            $this->assignContentToCache(
+                $summary,
+                $collection,
+                $cache_control,
+                $cache_summary
+            );
+        }
+
+        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
+    }
+
+    /**
+     * Return the categories summary for a resource
+     *
+     * @param int $resource_type_id
+     * @param int $resource_id
+     * @param array $parameters
+     *
+     * @return JsonResponse
+     */
+    private function categoriesSummary(
+        int $resource_type_id,
+        int $resource_id,
+        array $parameters
+    ): JsonResponse {
+        $cache_control = new Cache\Control($this->user_id);
+        $cache_control->setTtlOneWeek();
+
+        $cache_summary = new Cache\Summary();
+        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
+
+        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
             $summary = $this->model->categoriesSummary(
                 $resource_type_id,
+                $resource_id,
                 $parameters
             );
 
             $collection = array_map(
                 static function ($category) {
-                    return (new ResourceTypeItemCategoryTransformer($category))->asArray();
+                    return (new ItemCategoryTransformer($category))->asArray();
                 },
                 $summary
             );
@@ -582,10 +493,109 @@ class ResourceTypeItemController extends Controller
     }
 
     /**
-     * Return the total summary for all the resources in the resource type
-     * for a specific category
+     * Return the subcategories summary for a category
      *
      * @param int $resource_type_id
+     * @param int $resource_id
+     * @param int $category_id
+     * @param array $parameters
+     *
+     * @return JsonResponse
+     */
+    private function subcategoriesSummary(
+        int $resource_type_id,
+        int $resource_id,
+        int $category_id,
+        array $parameters
+    ): JsonResponse {
+        $cache_control = new Cache\Control($this->user_id);
+        $cache_control->setTtlOneWeek();
+
+        $cache_summary = new Cache\Summary();
+        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
+
+        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
+
+            $summary = $this->model->subCategoriesSummary(
+                $resource_type_id,
+                $resource_id,
+                $category_id,
+                $parameters
+            );
+
+            $collection = array_map(
+                static function ($subcategory) {
+                    return (new ItemSubcategoryTransformer($subcategory))->asArray();
+                },
+                $summary
+            );
+
+            $this->assignContentToCache(
+                $summary,
+                $collection,
+                $cache_control,
+                $cache_summary
+            );
+        }
+
+        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
+    }
+
+    /**
+     * Return the subcategories summary for a category
+     *
+     * @param int $resource_type_id
+     * @param int $resource_id
+     * @param int $category_id
+     * @param int $subcategory_id
+     * @param array $parameters
+     *
+     * @return JsonResponse
+     */
+    private function subcategorySummary(
+        int $resource_type_id,
+        int $resource_id,
+        int $category_id,
+        int $subcategory_id,
+        array $parameters
+    ): JsonResponse {
+        $cache_control = new Cache\Control($this->user_id);
+        $cache_control->setTtlOneWeek();
+
+        $cache_summary = new Cache\Summary();
+        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
+
+        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
+
+            $summary = $this->model->subCategorySummary(
+                $resource_type_id,
+                $resource_id,
+                $category_id,
+                $subcategory_id,
+                $parameters
+            );
+
+            $collection = [];
+            if (array_key_exists(0, $summary)) {
+                $collection = (new ItemSubcategoryTransformer($summary[0]))->asArray();
+            }
+
+            $this->assignContentToCache(
+                $summary,
+                $collection,
+                $cache_control,
+                $cache_summary
+            );
+        }
+
+        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
+    }
+
+    /**
+     * Return the category summary for a resource
+     *
+     * @param int $resource_type_id
+     * @param int $resource_id
      * @param int $category_id
      * @param array $parameters
      *
@@ -593,10 +603,10 @@ class ResourceTypeItemController extends Controller
      */
     private function categorySummary(
         int $resource_type_id,
+        int $resource_id,
         int $category_id,
         array $parameters
-    ): JsonResponse
-    {
+    ): JsonResponse {
         $cache_control = new Cache\Control($this->user_id);
         $cache_control->setTtlOneWeek();
 
@@ -607,13 +617,14 @@ class ResourceTypeItemController extends Controller
 
             $summary = $this->model->categorySummary(
                 $resource_type_id,
+                $resource_id,
                 $category_id,
                 $parameters
             );
 
             $collection = [];
             if (array_key_exists(0, $summary)) {
-                $collection = (new ResourceTypeItemCategoryTransformer($summary[0]))->asArray();
+                $collection = (new ItemCategoryTransformer($summary[0]))->asArray();
             }
 
             $this->assignContentToCache(
@@ -631,6 +642,7 @@ class ResourceTypeItemController extends Controller
      * Return a filtered summary
      *
      * @param int $resource_type_id
+     * @param int $resource_id
      * @param int|null $category_id
      * @param int|null $subcategory_id
      * @param int|null $year
@@ -643,6 +655,7 @@ class ResourceTypeItemController extends Controller
      */
     private function filteredSummary(
         int $resource_type_id,
+        int $resource_id,
         int $category_id = null,
         int $subcategory_id = null,
         int $year = null,
@@ -650,8 +663,7 @@ class ResourceTypeItemController extends Controller
         array $parameters = [],
         array $search_parameters = [],
         array $filter_parameters = []
-    ): JsonResponse
-    {
+    ): JsonResponse {
         $cache_control = new Cache\Control($this->user_id);
         $cache_control->setTtlOneWeek();
 
@@ -662,6 +674,7 @@ class ResourceTypeItemController extends Controller
 
             $summary = $this->model->filteredSummary(
                 $resource_type_id,
+                $resource_id,
                 $category_id,
                 $subcategory_id,
                 $year,
@@ -692,21 +705,19 @@ class ResourceTypeItemController extends Controller
     }
 
     /**
-     * Return the total summary for all the resources in the resource type
-     * and category grouped by subcategory
+     * Return the total summary for a resource, total cost of ownership
      *
      * @param int $resource_type_id
-     * @param int $category_id
+     * @param int $resource_id
      * @param array $parameters
      *
      * @return JsonResponse
      */
-    private function subcategoriesSummary(
+    private function tcoSummary(
         int $resource_type_id,
-        int $category_id,
+        int $resource_id,
         array $parameters
-    ): JsonResponse
-    {
+    ): JsonResponse {
         $cache_control = new Cache\Control($this->user_id);
         $cache_control->setTtlOneWeek();
 
@@ -715,68 +726,21 @@ class ResourceTypeItemController extends Controller
 
         if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
 
-            $summary = $this->model->subcategoriesSummary(
+            $summary = $this->model->summary(
                 $resource_type_id,
-                $category_id,
+                $resource_id,
                 $parameters
             );
 
-            $collection = array_map(
-                static function ($category) {
-                    return (new ResourceTypeItemSubcategoryTransformer($category))->asArray();
-                },
-                $summary
-            );
-
-            $this->assignContentToCache(
-                $summary,
-                $collection,
-                $cache_control,
-                $cache_summary
-            );
-        }
-
-        return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
-    }
-
-    /**
-     * Return the total summary for all the resources in the resource type
-     * for a specific category and subcategory
-     *
-     * @param int $resource_type_id
-     * @param int $category_id
-     * @param int $subcategory_id
-     * @param array $parameters
-     *
-     * @return JsonResponse
-     */
-    private function subcategorySummary(
-        int $resource_type_id,
-        int $category_id,
-        int $subcategory_id,
-        array $parameters
-    ): JsonResponse
-    {
-        $cache_control = new Cache\Control($this->user_id);
-        $cache_control->setTtlOneWeek();
-
-        $cache_summary = new Cache\Summary();
-        $cache_summary->setFromCache($cache_control->get(request()->getRequestUri()));
-
-        if ($cache_control->cacheable() === false || $cache_summary->valid() === false) {
-
-            $summary = $this->model->subcategorySummary(
-                $resource_type_id,
-                $category_id,
-                $subcategory_id,
-                $parameters
-            );
-
-            $collection = [];
-            if (array_key_exists(0, $summary)) {
-                $collection = (new ResourceTypeItemSubcategoryTransformer($summary[0]))->asArray();
+            $total = '0.00';
+            if (array_key_exists(0, $summary) && array_key_exists('total', $summary[0])) {
+                $total = number_format($summary[0]['total'], 2, '.', '');
             }
 
+            $collection = [
+                'total' => $total
+            ];
+
             $this->assignContentToCache(
                 $summary,
                 $collection,
@@ -788,36 +752,33 @@ class ResourceTypeItemController extends Controller
         return response()->json($cache_summary->collection(), 200, $cache_summary->headers());
     }
 
-
     /**
-     * Generate the OPTIONS request for items summary route
+     * Generate the OPTIONS request for items route
      *
      * @param string $resource_type_id
+     * @param string $resource_id
      *
      * @return JsonResponse
-     *
      */
-    public function optionsIndex(string $resource_type_id): JsonResponse
+    public function optionsIndex(string $resource_type_id, string $resource_id): JsonResponse
     {
-        Route\Validate::resourceType(
-            (int) $resource_type_id,
+        Route\Validate::resource(
+            $resource_type_id,
+            $resource_id,
             $this->permitted_resource_types
         );
 
         $item_interface = Factory::summaryItem($resource_type_id);
 
-        $permissions = Route\Permission::resourceType(
+        $permissions = Route\Permission::resource(
             $resource_type_id,
+            $resource_id,
             $this->permitted_resource_types
         );
 
-        $get = Get::init()->
-            setSearchable($item_interface->searchParametersConfig())->
-            setParameters($item_interface->collectionParametersConfig())->
-            setFilterable($item_interface->filterParametersConfig())->
-            setDescription('route-descriptions.summary-resource-type-item-GET-index')->
-            setAuthenticationStatus($permissions['view'])->
-            option();
+        $get = Get::init()->setParameters($item_interface->collectionParametersConfig())->setSearchable($item_interface->searchParametersConfig())
+            ->setFilterable($item_interface->filterParametersConfig())->setDescription('route-descriptions.summary_GET_resource-type_resource_items')
+            ->setAuthenticationStatus($permissions['view'])->option();
 
         return $this->optionsResponse($get, 200);
     }
