+++ app/Http/Controllers/Summary/ResourceTypeView.php
@@ -18,7 +18,7 @@ use Illuminate\Support\Facades\Config;
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceTypeController extends Controller
+class ResourceTypeView extends Controller
 {
     /**
      * Return a summary of the resource types
